requirements.txt
cycler==0.11.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.3
Pillow==8.4.0
PyOpenGL==3.1.5
pyparsing==3.0.4
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
six==1.16.0
